dnfshellrpmupgrade => dnfupgradetransaction
Additionally made to use the new code from the dnfplugin library.
